refactor: Extract audio integrity checks to a dedicated module
Move audio stream integrity verification logic from `ts2mp4.py` to a new module `audio_integrity.py`. This improves modularity and testability.

- New module `ts2mp4/audio_integrity.py` contains `_get_audio_stream_count`, `_get_audio_stream_md5`, and `verify_audio_stream_integrity`.
- Corresponding tests are moved from `tests/test_ts2mp4.py` to `tests/test_audio_integrity.py`.
- `pytest-mock` is added as a development dependency for testing purposes.
- `ts2mp4.py` now imports `verify_audio_stream_integrity` from the new module.
