+++ pyproject.toml
@@ -18,6 +18,7 @@ flake8 = "^7.3.0"
 mypy = "*"
 pydocstyle = "*"
 pytest = "^8.4.1"
+pytest-mock = "^3.14.1"
 
 [tool.poetry.scripts]
 ts2mp4 = "ts2mp4.cli:app"
