--- tests/test_ts2mp4.py
@@ -1,17 +1,0 @@
-from pathlib import Path
-
-from ts2mp4.ts2mp4 import _get_audio_stream_count, _get_audio_stream_md5
-
-
-def test_get_audio_stream_count(ts_file: Path):
-    """Test the _get_audio_stream_count helper function."""
-    expected_stream_count = 1
-    actual_stream_count = _get_audio_stream_count(ts_file)
-    assert actual_stream_count == expected_stream_count
-
-
-def test_get_audio_stream_md5(ts_file: Path):
-    """Test the _get_audio_stream_md5 helper function."""
-    expected_md5 = "9db9dd4cb46b9678894578946158955b"
-    actual_md5 = _get_audio_stream_md5(ts_file, 0)
-    assert actual_md5 == expected_md5
